Create WaterLevel&TemperatureControl
A pressure sensor connected to the fog chamber reads the pressure continuously and compares it to the average of first 20 readings. If the pressure is lower than 1mm of water it actuates a solenoid valve until it reaches the average value. Simultaneously, temperature, relative humidity and atmospheric pressure are recorded using a BME280.